fix type cast warning

+++ mxnet_cpp_object_detection/object_detector_filter.hpp
@@ -96,7 +96,7 @@ public:
 
     struct result_type
     {
-        double confidence_ = 0.0;
+        float confidence_ = 0.0f;
         item_type item_;
         cv::Rect roi_;
     };
@@ -104,12 +104,12 @@ public:
     explicit object_detector_filter(std::vector<item_type> const &items_to_detect,
                                     cv::Size const &obj_detector_process_size,
                                     cv::Size const &image_size,
-                                    double min_confidence = 0.4);
+                                    float min_confidence = 0.4f);
 
     std::vector<result_type> filter(std::vector<mxnet::cpp::NDArray> const &input) const;
 
     void set_image_size(cv::Size const &input) noexcept;
-    void set_min_confidence(double input) noexcept;
+    void set_min_confidence(float input) noexcept;
     void set_items_to_detect(std::vector<item_type> const &items_to_detect);
     void set_obj_detector_process_size(cv::Size const &obj_detector_process_size) noexcept;
 
@@ -117,7 +117,7 @@ private:
     cv::Rect normalize_points(float x1, float y1, float x2, float y2) const noexcept;
 
     cv::Size image_size_;
-    double min_confidence_;
+    float min_confidence_;
     std::deque<bool> items_to_detect_;
     cv::Size obj_detector_process_size_;
 };
